+++ src/view/console.rs
@@ -33,7 +33,7 @@ impl Renderer for Console<'_, '_> {
                     " needed: [{}]",
                     term.inference
                         .requirements()
-                        .map(|x| &x.task.goal.term)
+                        .map(|x| &x.task.goal().term)
                         .format(", ")
                 )
             }